Read configuration file granularly (#160)
