+++ pylama.ini
@@ -13,3 +13,6 @@ disable = R
 
 [pylama:*/packtPublishingFreeEbook.py]
 linters = pycodestyle,pyflakes
+
+[pylama:*/downloader.py]
+linters=pycodestyle,pyflakes
